move screenshot utils from utils to shell
